requirements.txt
flask
flask-cors
gunicorn
supabase
openai
requests
python-dotenv
pydub
pytest
